Making the d3 graph more re-usable, so we can use it on other reports
