+++ MrCMS.Web/Areas/Admin/Models/LineGraphData.cs
@@ -1,0 +1,10 @@
+namespace MrCMS.Web.Areas.Admin.Models
+{
+    public class LineGraphData
+    {
+        // ReSharper disable InconsistentNaming - for use in JS
+        public object x { get; set; }
+        public object y { get; set; }
+        // ReSharper restore InconsistentNaming
+    }
+}
