+++ MrCMS.Web/Areas/Admin/Services/UserSubscriptionReports/UserSubscriptionReportsService.cs
@@ -9,12 +9,12 @@ using MrCMS.Paging;
 using MrCMS.Services.Events;
 using MrCMS.Services.Events.Args;
 using MrCMS.Settings;
+using MrCMS.Web.Areas.Admin.Models;
 using MrCMS.Website;
 using NHibernate;
 using NHibernate.Linq;
 using NHibernate.Criterion;
 using System.Linq;
-using MrCMS.Web.Areas.Admin.Helpers.D3Reports;
 using MrCMS.Web.Areas.Admin.Models.UserSubscriptionReports;
 
 namespace MrCMS.Web.Areas.Admin.Services.UserSubscriptionReports
@@ -28,11 +28,13 @@ namespace MrCMS.Web.Areas.Admin.Services.UserSubscriptionReports
             _session = session;
         }
 
-        public IEnumerable<object> GetAllSubscriptions(UserSubscriptionReportsSearchQuery searchQuery)
+        public IEnumerable<LineGraphData> GetAllSubscriptions(UserSubscriptionReportsSearchQuery searchQuery)
         {
-            return _session.Query<User>().Where(c => c.CreatedOn >= searchQuery.StartDate && c.CreatedOn <= searchQuery.EndDate).AsEnumerable()
-                           .GroupBy(c => c.CreatedOn.ToString("MMM") + " " + c.CreatedOn.Year)
-                           .Select(c => new { JoiningMonthYear = c.Key, Count = c.Count() });
+            return _session.Query<User>()
+                .Where(c => c.CreatedOn >= searchQuery.StartDate && c.CreatedOn <= searchQuery.EndDate)
+                .AsEnumerable()
+                .GroupBy(c => c.CreatedOn.ToString("MMM") + " " + c.CreatedOn.Year)
+                .Select(c => new LineGraphData {x = c.Key, y = c.Count()}).ToList();
         }
     }
 }
